+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## v5.33.0 - 2022-06-10
+
+### Changed
+
+- Spinner Upgrade Tech changes targeting instead of rate of fire
+
+### Fixed
+
+- Skitter weapon not receiving correct stats or buffs
+- Firefly weapon not receiving correct stats or buffs
+
 ## v5.32.0 - 2022-06-09
 
 ### Changed